distributed-tracing/app/requirements.txt
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-instrumentation-flask==0.45b0
opentelemetry-instrumentation-requests==0.45b0
opentelemetry-instrumentation-logging==0.45b0

azure-monitor-opentelemetry==1.4.2

flask==3.0.3
python-dotenv==1.0.1